# backend_fashion_app/internal/common/httpx/docs/swagger.yaml
basePath: /api/v1
definitions:
  CancelOrderRequest:
    properties:
      reason:
        example: Đặt nhầm sản phẩm
        type: string
    required:
    - reason
    type: object
  CreateOrderRequest:
    description: Request body for creating a new order from cart items
    properties:
      address_id:
        example: 1
        type: integer
      cart_item_ids:
        example:
        - 1
        - 2
        - 3
        items:
          type: integer
        minItems: 1
        type: array
      note:
        example: Giao hàng giờ hành chính
        type: string
      payment_method:
        enum:
        - cod
        - bank_transfer
        - e_wallet
        example: cod
        type: string
      promotion_codes:
        items:
          type: string
        type: array
    required:
    - address_id
    - cart_item_ids
    - payment_method
    type: object
  OrderItemResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      price:
        example: 150000
        type: number
      product_id:
        example: 1
        type: integer
      product_image:
        example: https://example.com/image.jpg
        type: string
      product_name:
        example: Áo thun nam basic
        type: string
      quantity:
        example: 2
        type: integer
      subtotal:
        example: 300000
        type: number
    type: object
  OrderListResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      orders:
        items:
          $ref: '#/definitions/OrderResponse'
        type: array
      total:
        example: 100
        type: integer
    type: object
  OrderResponse:
    properties:
      cancel_reason:
        example: ""
        type: string
      cancelled_at:
        example: ""
        type: string
      confirmed_at:
        example: "2024-01-21T10:30:00Z"
        type: string
      created_at:
        example: "2024-01-21T10:00:00Z"
        type: string
      customer_id:
        example: 1
        type: integer
      delivered_at:
        example: "2024-01-22T10:00:00Z"
        type: string
      discount_amount:
        example: 50000
        type: number
      final_amount:
        example: 480000
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      items:
        items:
          $ref: '#/definitions/OrderItemResponse'
        type: array
      note:
        example: Giao hàng giờ hành chính
        type: string
      order_code:
        example: 1731570000123456789
        type: integer
      order_number:
        example: ORD-20240121-abc123
        type: string
      payment_method:
        example: cod
        type: string
      payment_status:
        example: pending
        type: string
      shipped_at:
        example: "2024-01-21T14:00:00Z"
        type: string
      shipping_address:
        example: 123 Đường ABC
        type: string
      shipping_district:
        example: Quận 1
        type: string
      shipping_fee:
        example: 30000
        type: number
      shipping_name:
        example: Nguyễn Văn A
        type: string
      shipping_phone:
        example: "0901234567"
        type: string
      shipping_province:
        example: Hồ Chí Minh
        type: string
      shipping_ward:
        example: Phường Bến Nghé
        type: string
      shop_id:
        example: 2
        type: integer
      status:
        example: pending
        type: string
      total_amount:
        example: 500000
        type: number
      updated_at:
        example: "2024-01-21T10:00:00Z"
        type: string
    type: object
  UpdateOrderStatusRequest:
    properties:
      status:
        enum:
        - confirmed
        - processing
        - shipping
        - delivered
        - cancelled
        - returned
        example: confirmed
        type: string
    required:
    - status
    type: object
...
  resp.Envelope:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: Auth (local, Google) + Profile (customer/shop) + Catalog
  title: MyFashion API
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Lấy danh sách đơn hàng của tôi (customer hoặc shop)
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: Filter by status
        enum:
        - pending
        - confirmed
        - processing
        - shipping
        - delivered
        - cancelled
        - returned
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Get My Orders
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: 'Tạo đơn hàng mới từ giỏ hàng (customer), các loại thanh toán:
        cod, bank_transfer, e_wallet'
      parameters:
      - description: Order info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Create Order
      tags:
      - Order
  /orders/{id}:
    get:
      description: Lấy thông tin đơn hàng theo ID
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Get Order by ID
      tags:
      - Order
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Hủy đơn hàng (customer hoặc shop)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Cancel reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Cancel Order
      tags:
      - Order
  /orders/{id}/confirm:
    post:
      description: Xác nhận đơn hàng (shop only)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Confirm Order
      tags:
      - Order
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Cập nhật trạng thái đơn hàng (shop only)
      parameters:
      - description: Order ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resp.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/resp.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Envelope'
      security:
      - Bearer: []
      summary: Update Order Status
      tags:
      - Order
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"